requirements.txt
pandas
requests
python-dotenv
metaapi-cloud-sdk
django